--- src/main/java/asw/InciManagerApplication.java
@@ -1,14 +1,0 @@
-package asw; 
-
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class InciManagerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(InciManagerApplication.class, args);
-	}
-
-}
