--- src/main/java/asw/inci_manager/inci_manager_gest/repositories/OperarioRepository.java
@@ -1,7 +1,0 @@
-package asw.inci_manager.inci_manager_gest.repositories;
-
-import asw.inci_manager.inci_manager_gest.entities.Operario;
-import org.springframework.data.repository.CrudRepository;
-
-public interface OperarioRepository extends CrudRepository<Operario,Long> {
-}
