--- src/main/java/asw/inci_manager/inci_manager_gest/responses/RespuestaREST.java
@@ -1,5 +1,0 @@
-package asw.inci_manager.inci_manager_gest.responses;
-
-public interface RespuestaREST {
-
-}
