--- src/main/java/asw/inci_manager/inci_manager_gest/services/AgentService.java
@@ -1,17 +1,0 @@
-package asw.inci_manager.inci_manager_gest.services;
-
-import asw.inci_manager.inci_manager_gest.entities.Agent;
-import asw.inci_manager.inci_manager_gest.repositories.AgentRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class AgentService {
-
-    @Autowired
-    private AgentRepository agentRepository;
-
-    public Agent getAgentByEmailFlexible(String email){
-        return agentRepository.findByEmailFlexible(email);
-    }
-}
