--- src/main/java/asw/inci_manager/util/Estado.java
@@ -1,5 +1,0 @@
-package asw.inci_manager.util;
-
-public enum Estado {
-    ABIERTA, EN_PROCESO, CERRADA, ANULADA
-}
