+++ src/main/java/es/uniovi/asw/e3b/InciManagerApplication.java
@@ -1,0 +1,14 @@
+package es.uniovi.asw.e3b;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class InciManagerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(InciManagerApplication.class, args);
+	}
+
+}
