+++ src/main/java/es/uniovi/asw/e3b/inci_manager/inci_manager_gest/repositories/OperarioRepository.java
@@ -1,0 +1,7 @@
+package es.uniovi.asw.e3b.inci_manager.inci_manager_gest.repositories;
+
+import es.uniovi.asw.e3b.inci_manager.inci_manager_gest.entities.Operario;
+import org.springframework.data.repository.CrudRepository;
+
+public interface OperarioRepository extends CrudRepository<Operario,Long> {
+}
