+++ src/main/java/es/uniovi/asw/e3b/inci_manager/inci_manager_gest/responses/RespuestaREST.java
@@ -1,0 +1,5 @@
+package es.uniovi.asw.e3b.inci_manager.inci_manager_gest.responses;
+
+public interface RespuestaREST {
+
+}
