+++ src/main/java/es/uniovi/asw/e3b/inci_manager/inci_manager_gest/services/AgentService.java
@@ -1,0 +1,17 @@
+package es.uniovi.asw.e3b.inci_manager.inci_manager_gest.services;
+
+import es.uniovi.asw.e3b.inci_manager.inci_manager_gest.entities.Agent;
+import es.uniovi.asw.e3b.inci_manager.inci_manager_gest.repositories.AgentRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class AgentService {
+
+    @Autowired
+    private AgentRepository agentRepository;
+
+    public Agent getAgentByEmailFlexible(String email){
+        return agentRepository.findByEmailFlexible(email);
+    }
+}
