+++ src/main/java/es/uniovi/asw/e3b/inci_manager/util/Estado.java
@@ -1,0 +1,5 @@
+package es.uniovi.asw.e3b.inci_manager.util;
+
+public enum Estado {
+    ABIERTA, EN_PROCESO, CERRADA, ANULADA
+}
