--- src/test/java/asw/inci_manager/CucumberTest.java
@@ -1,11 +1,0 @@
-package asw.inci_manager;
-
-import org.junit.runner.RunWith;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(features = "src/test/resources/features", glue = "asw.inci_manager.steps")
-public class CucumberTest{
-}
