+++ src/test/java/es/uniovi/asw/e3b/inci_manager/InciManagerapplicationTestLogueado.java
@@ -1,0 +1,68 @@
+package es.uniovi.asw.e3b.inci_manager;
+
+import es.uniovi.asw.e3b.InciManagerApplication;
+import es.uniovi.asw.e3b.inci_manager.inci_manager_gest.entities.Agent;
+import org.junit.Before;
+import org.junit.FixMethodOrder;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.MethodSorters;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.IntegrationTest;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.mock.web.MockHttpSession;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.ResultMatcher;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import javax.servlet.Filter;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+
+@SuppressWarnings("deprecation")
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringBootTest(classes = InciManagerApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ContextConfiguration
+@IntegrationTest({ "server.port=0" })
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+public class InciManagerapplicationTestLogueado {
+
+    @Autowired
+    private WebApplicationContext context;
+
+    @Autowired
+    private Filter springSecurityFilterChain;
+
+    private MockMvc mvc;
+
+
+    @Before
+    public void setup() {
+        mvc = MockMvcBuilders
+                .webAppContextSetup(context)
+                .addFilters(springSecurityFilterChain)
+                .build();
+    }
+
+    @Test
+    public void testHomeLogueado() throws Exception {
+        mvc.perform(get("/"));
+    }
+
+    @Test
+    public void testHomeMockup() throws Exception {
+        mvc.perform(get("/home"));
+    }
+
+    @Test
+    public void testLoginMockup() throws Exception {
+        mvc.perform(get("/login"));
+    }
+
+}
+
